+++ spec/ax-command-bar-widget.spec.js
@@ -5,42 +5,71 @@
  */
 define( [
    'json!../widget.json',
-   '../ax-command-bar-widget',
-   'laxar/laxar_testing',
+   'laxar-mocks',
+   'laxar',
+   'angular-mocks',
+   './helpers/spec_helper',
    './fixtures'
-], function( descriptor, navigator, ax, testData ) {
+], function( descriptor, axMocks, ax, ngMocks, buttonMatchers, testData ) {
    'use strict';
 
-   var testBed_;
+   var widgetEventBus;
+   var widgetScope;
+   var testEventBus;
+
+   ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+   function createSetup( widgetConfiguration ) {
+
+      beforeEach( axMocks.createSetupForWidget( descriptor, {
+         knownMissingResources: [ 'ax-button-list-control.css' ]
+      } ) );
+
+      beforeEach( function() {
+         axMocks.widget.configure( widgetConfiguration );
+      } );
+
+      beforeEach( axMocks.widget.load );
+
+      beforeEach( function() {
+         widgetScope = axMocks.widget.$scope;
+         widgetEventBus = axMocks.widget.axEventBus;
+         testEventBus = axMocks.eventBus;
+      } );
+   }
+
+   ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+   beforeEach( function() {
+      jasmine.addMatchers( buttonMatchers );
+   } );
+
+   ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+   afterEach( function() {
+      axMocks.tearDown();
+   } );
 
    ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 
    describe( 'An ax-command-bar-widget', function() {
 
-      beforeEach( function setup() {
-         testBed_ = testBedAfterDidNavigate( this, { buttons: [] } );
-      } );
-
-      ////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-      afterEach( function() {
-         testBed_.tearDown();
-      } );
+      createSetup( { buttons: [] } );
 
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       it( 'defines three distinct areas for buttons (R1.1)', function() {
-         expect( testBed_.scope.model.areas.left ).toBeDefined();
-         expect( testBed_.scope.model.areas.center ).toBeDefined();
-         expect( testBed_.scope.model.areas.right ).toBeDefined();
+         expect( widgetScope.model.areas.left ).toBeDefined();
+         expect( widgetScope.model.areas.center ).toBeDefined();
+         expect( widgetScope.model.areas.right ).toBeDefined();
       } );
 
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       it( 'renders no buttons by default', function() {
-         expect( testBed_.scope.model.areas.left.length ).toBe( 0 );
-         expect( testBed_.scope.model.areas.center.length ).toBe( 0 );
-         expect( testBed_.scope.model.areas.right.length ).toBe( 0 );
+         expect( widgetScope.model.areas.left.length ).toBe( 0 );
+         expect( widgetScope.model.areas.center.length ).toBe( 0 );
+         expect( widgetScope.model.areas.right.length ).toBe( 0 );
       } );
 
    } );
@@ -51,38 +80,36 @@ define( [
 
       describe( 'when no buttons are disabled', function() {
 
-         beforeEach( function() {
-            testBed_ = testBedAfterDidNavigate( this, testData.customAllEnabledButtons );
-         } );
+         createSetup( testData.customAllEnabledButtons );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'puts them into the correct areas (R1.2, R2.4)', function() {
-            expect( testBed_.scope.model.areas.left ).toContainButtonWithAction( 'action1' );
-            expect( testBed_.scope.model.areas.left ).toContainButtonWithAction( 'action2' );
-            expect( testBed_.scope.model.areas.left ).toContainButtonWithAction( 'action3' );
-            expect( testBed_.scope.model.areas.center ).toContainButtonWithAction( 'action4' );
-            expect( testBed_.scope.model.areas.right ).toContainButtonWithAction( 'action5' );
+            expect( widgetScope.model.areas.left ).toContainButtonWithAction( 'action1' );
+            expect( widgetScope.model.areas.left ).toContainButtonWithAction( 'action2' );
+            expect( widgetScope.model.areas.left ).toContainButtonWithAction( 'action3' );
+            expect( widgetScope.model.areas.center ).toContainButtonWithAction( 'action4' );
+            expect( widgetScope.model.areas.right ).toContainButtonWithAction( 'action5' );
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'sorts them by their index (R2.4)', function() {
-            expect( testBed_.scope.model.areas.left[0] ).toHaveAction( 'action3' );
-            expect( testBed_.scope.model.areas.left[1] ).toHaveAction( 'action1' );
-            expect( testBed_.scope.model.areas.left[2] ).toHaveAction( 'action2' );
+            expect( widgetScope.model.areas.left[ 0 ] ).toHaveAction( 'action3' );
+            expect( widgetScope.model.areas.left[ 1 ] ).toHaveAction( 'action1' );
+            expect( widgetScope.model.areas.left[ 2 ] ).toHaveAction( 'action2' );
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'sets the css class for class-attribute (R2.5)', function() {
-            expect( testBed_.scope.model.areas.right[0].classes[ 'btn-success' ] ).toBe( true );
+            expect( widgetScope.model.areas.right[ 0 ].classes[ 'btn-success' ] ).toBe( true );
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'sets the css class for size-attribute (R2.6)', function() {
-            expect( testBed_.scope.model.areas.center[0].classes[ 'btn-sm' ] ).toBe( true );
+            expect( widgetScope.model.areas.center[ 0 ].classes[ 'btn-sm' ] ).toBe( true );
          } );
 
       } );
@@ -92,26 +119,27 @@ define( [
       describe( 'when some buttons are disabled', function() {
 
          beforeEach( function() {
-            testData.customAllEnabledButtons.buttons[1].enabled = false;
-            testData.customAllEnabledButtons.buttons[3].enabled = false;
-            testBed_ = testBedAfterDidNavigate( this, testData.customAllEnabledButtons );
+            testData.customAllEnabledButtons.buttons[ 1 ].enabled = false;
+            testData.customAllEnabledButtons.buttons[ 3 ].enabled = false;
          } );
+
+         createSetup( testData.customAllEnabledButtons );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          afterEach( function() {
-            testData.customAllEnabledButtons.buttons[1].enabled = true;
-            testData.customAllEnabledButtons.buttons[3].enabled = true;
+            testData.customAllEnabledButtons.buttons[ 1 ].enabled = true;
+            testData.customAllEnabledButtons.buttons[ 3 ].enabled = true;
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'does not render these buttons (R2.1)', function() {
-            expect( testBed_.scope.model.areas.left ).toContainButtonWithAction( 'action1' );
-            expect( testBed_.scope.model.areas.left ).not.toContainButtonWithAction( 'action2' );
-            expect( testBed_.scope.model.areas.left ).toContainButtonWithAction( 'action3' );
-            expect( testBed_.scope.model.areas.center ).not.toContainButtonWithAction( 'action4' );
-            expect( testBed_.scope.model.areas.right ).toContainButtonWithAction( 'action5' );
+            expect( widgetScope.model.areas.left ).toContainButtonWithAction( 'action1' );
+            expect( widgetScope.model.areas.left ).not.toContainButtonWithAction( 'action2' );
+            expect( widgetScope.model.areas.left ).toContainButtonWithAction( 'action3' );
+            expect( widgetScope.model.areas.center ).not.toContainButtonWithAction( 'action4' );
+            expect( widgetScope.model.areas.right ).toContainButtonWithAction( 'action5' );
          } );
 
       } );
@@ -120,15 +148,13 @@ define( [
 
       describe( 'when no buttons are disabled and some buttons have index and some have no index', function() {
 
-         beforeEach( function() {
-            testBed_ = testBedAfterDidNavigate( this, testData.sortTestButtons.customButtons );
-         } );
+         createSetup( testData.sortTestButtons.customButtons  );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'sorts them by their index (R2.4)', function() {
             var buttonOrder = [ 15, 1, 2, 3, 4, 5, 10, 11, 6, 12, 13, 7, 14, 8, 9 ];
-            testBed_.scope.model.areas.left.forEach( function( button, i ) {
+            widgetScope.model.areas.left.forEach( function( button, i ) {
                expect( button ).toHaveAction( 'action' + buttonOrder[ i ] );
             } );
          } );
@@ -143,24 +169,22 @@ define( [
 
       describe( 'when no buttons are disabled', function() {
 
-         beforeEach( function() {
-            testBed_ = testBedAfterDidNavigate( this, testData.defaultAllEnabledButtons );
-         } );
+         createSetup( testData.defaultAllEnabledButtons );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'puts them into the correct areas (R3.1)', function() {
-            expect( testBed_.scope.model.areas.left ).toContainButtonWithAction( 'previous' );
-            expect( testBed_.scope.model.areas.center ).toContainButtonWithAction( 'help' );
-            expect( testBed_.scope.model.areas.right ).toContainButtonWithAction( 'next' );
-            expect( testBed_.scope.model.areas.right ).toContainButtonWithAction( 'cancel' );
+            expect( widgetScope.model.areas.left ).toContainButtonWithAction( 'previous' );
+            expect( widgetScope.model.areas.center ).toContainButtonWithAction( 'help' );
+            expect( widgetScope.model.areas.right ).toContainButtonWithAction( 'next' );
+            expect( widgetScope.model.areas.right ).toContainButtonWithAction( 'cancel' );
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'sorts them by their index (R3.1)', function() {
-            expect( testBed_.scope.model.areas.right[0] ).toHaveAction( 'cancel' );
-            expect( testBed_.scope.model.areas.right[1] ).toHaveAction( 'next' );
+            expect( widgetScope.model.areas.right[ 0 ] ).toHaveAction( 'cancel' );
+            expect( widgetScope.model.areas.right[ 1 ] ).toHaveAction( 'next' );
          } );
 
       } );
@@ -172,8 +196,9 @@ define( [
          beforeEach( function() {
             testData.defaultAllEnabledButtons.previous.enabled = false;
             testData.defaultAllEnabledButtons.next.enabled = false;
-            testBed_ = testBedAfterDidNavigate( this, testData.defaultAllEnabledButtons );
          } );
+
+         createSetup( testData.defaultAllEnabledButtons );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -185,10 +210,10 @@ define( [
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'does not render these buttons (R3.1)', function() {
-            expect( testBed_.scope.model.areas.left ).not.toContainButtonWithAction( 'previous' );
-            expect( testBed_.scope.model.areas.center ).toContainButtonWithAction( 'help' );
-            expect( testBed_.scope.model.areas.right ).not.toContainButtonWithAction( 'next' );
-            expect( testBed_.scope.model.areas.right ).toContainButtonWithAction( 'cancel' );
+            expect( widgetScope.model.areas.left ).not.toContainButtonWithAction( 'previous' );
+            expect( widgetScope.model.areas.center ).toContainButtonWithAction( 'help' );
+            expect( widgetScope.model.areas.right ).not.toContainButtonWithAction( 'next' );
+            expect( widgetScope.model.areas.right ).toContainButtonWithAction( 'cancel' );
          } );
 
       } );
@@ -197,15 +222,13 @@ define( [
 
       describe( 'and some of the buttons have index and some have no index', function() {
 
-         beforeEach( function() {
-            testBed_ = testBedAfterDidNavigate( this, testData.sortTestButtons.defaultButtons );
-         } );
+         createSetup( testData.sortTestButtons.defaultButtons );
 
          //////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'sorts them by their index and inserts default buttons before custom buttons having the same index (R2.4)', function() {
             var buttonOrder = [ 11, 7, 8, 1, 3, 4, 5, 9, 12, 10, 2 ];
-            testBed_.scope.model.areas.left.forEach( function( button, i ) {
+            widgetScope.model.areas.left.forEach( function( button, i ) {
                expect( button ).toHaveAction( 'action' + buttonOrder[ i ] );
             } );
          } );
@@ -219,42 +242,30 @@ define( [
 
       describe( '', function() {
 
-         var originalCustomButtons;
+         var features = ax.object.deepClone( testData.defaultAllEnabledButtons );
+         features.buttons = testData.customAllEnabledButtons.buttons;
 
-         beforeEach( function() {
-            originalCustomButtons = testData.defaultAllEnabledButtons.buttons;
-            var features = testData.defaultAllEnabledButtons;
-            features.buttons = testData.customAllEnabledButtons.buttons;
-
-            testBed_ = testBedAfterDidNavigate( this, features );
-            changeLocale( 'en-US' );
-         } );
-
-         /////////////////////////////////////////////////////////////////////////////////////////////////////
-
-         afterEach( function() {
-            testData.defaultAllEnabledButtons.buttons = originalCustomButtons;
-         } );
+         createSetup( features );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'inserts default buttons before custom buttons having the same index (R3.2)', function() {
-            expect( testBed_.scope.model.areas.right[0] ).toHaveAction( 'cancel' );
-            expect( testBed_.scope.model.areas.right[1] ).toHaveAction( 'action5' );
-            expect( testBed_.scope.model.areas.right[2] ).toHaveAction( 'next' );
+            expect( widgetScope.model.areas.right[ 0 ] ).toHaveAction( 'cancel' );
+            expect( widgetScope.model.areas.right[ 1 ] ).toHaveAction( 'action5' );
+            expect( widgetScope.model.areas.right[ 2 ] ).toHaveAction( 'next' );
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'creates a view label for a simple string (R2.2)', function() {
-            expect( testBed_.scope.model.areas.right[1].htmlLabel ).toEqual( 'Action 5' );
+            expect( widgetScope.model.areas.right[ 1 ].htmlLabel ).toEqual( 'Action 5' );
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'creates a view label for a localized object based on the current locale (R2.2)', function() {
             changeLocale( 'de' );
-            var next = testBed_.scope.model.areas.right[2];
+            var next = widgetScope.model.areas.right[ 2 ];
             expect( next.htmlLabel ).toEqual( '<i class=\"icon-circle-arrow-right\"></i> Weiter' );
          } );
 
@@ -264,15 +275,13 @@ define( [
 
          describe( 'and some of the buttons have index and some have no index', function() {
 
-            beforeEach( function() {
-               testBed_ = testBedAfterDidNavigate( this, testData.sortTestButtons.defaultAndCustomButtons );
-            } );
+            createSetup( testData.sortTestButtons.defaultAndCustomButtons );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'sorts them by their index and inserts default buttons before custom buttons having the same index (R2.4, R3.2)', function() {
                var buttonOrder = [ 16, 15, 1, 2, 3, 4, 5, 10, 11, 18, 6, 12, 13, 7, 14, 17, 8, 9 ];
-               testBed_.scope.model.areas.left.forEach( function( button, i ) {
+               widgetScope.model.areas.left.forEach( function( button, i ) {
                   expect( button ).toHaveAction( 'action' + buttonOrder[ i ] );
                } );
             } );
@@ -284,9 +293,7 @@ define( [
 
    describe( 'An AxCommandBarWidget with buttons', function() {
 
-      beforeEach( function() {
-         testBed_ = testBedAfterDidNavigate( this, testData.defaultAllEnabledButtons );
-      } );
+      createSetup( testData.defaultAllEnabledButtons );
 
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -298,22 +305,22 @@ define( [
          var previousButton;
 
          function publishFlagChange( flag, state ) {
-            testBed_.scope.eventBus.publish( 'didChangeFlag.' + flag + '.' + state, {
+            widgetScope.eventBus.publish( 'didChangeFlag.' + flag + '.' + state, {
                flag: flag,
                state: state
             } );
-            jasmine.Clock.tick( 0 );
+            testEventBus.flush();
          }
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          beforeEach( function() {
             mySpy = jasmine.createSpy( 'takeActionRequestSpy' );
-            testBed_.scope.eventBus.subscribe( 'takeActionRequest', mySpy );
+            widgetEventBus.subscribe( 'takeActionRequest', mySpy );
 
-            nextButton = testBed_.scope.model.areas.right[1];
-            helpButton = testBed_.scope.model.areas.center[0];
-            previousButton = testBed_.scope.model.areas.left[0];
+            nextButton = widgetScope.model.areas.right[ 1 ];
+            helpButton = widgetScope.model.areas.center[ 0 ];
+            previousButton = widgetScope.model.areas.left[ 0 ];
 
             publishFlagChange( 'helpAvailable', true );
          } );
@@ -351,31 +358,31 @@ define( [
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'on false the according css classes are removed (R2.7)', function() {
-               expect( nextButton.classes['ax-invisible'] ).toBe( true );
-               expect( nextButton.classes['ax-busy'] ).toBe( true );
-               expect( helpButton.classes['ax-omitted'] ).toBe( true );
-               expect( previousButton.classes['ax-disabled'] ).toBe( true );
+               expect( nextButton.classes[ 'ax-invisible' ] ).toBe( true );
+               expect( nextButton.classes[ 'ax-busy' ] ).toBe( true );
+               expect( helpButton.classes[ 'ax-omitted' ] ).toBe( true );
+               expect( previousButton.classes[ 'ax-disabled' ] ).toBe( true );
 
                publishFlagChange( 'guestUser', false );
                publishFlagChange( 'navigation', false );
                publishFlagChange( 'helpAvailable', true );
                publishFlagChange( 'notUndoable', false );
 
-               expect( nextButton.classes['ax-invisible'] ).toBe( false );
-               expect( nextButton.classes['ax-busy'] ).toBe( false );
-               expect( helpButton.classes['ax-omitted'] ).toBe( false );
-               expect( previousButton.classes['ax-disabled'] ).toBe( false );
+               expect( nextButton.classes[ 'ax-invisible' ] ).toBe( false );
+               expect( nextButton.classes[ 'ax-busy' ] ).toBe( false );
+               expect( helpButton.classes[ 'ax-omitted' ] ).toBe( false );
+               expect( previousButton.classes[ 'ax-disabled' ] ).toBe( false );
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
 
             it( 'no user interaction is possible (R2.7)', function() {
-               testBed_.scope.handleButtonClicked( nextButton );
-               testBed_.scope.handleButtonClicked( helpButton );
-               testBed_.scope.handleButtonClicked( previousButton );
-               jasmine.Clock.tick( 0 );
+               widgetScope.handleButtonClicked( nextButton );
+               widgetScope.handleButtonClicked( helpButton );
+               widgetScope.handleButtonClicked( previousButton );
+               testEventBus.flush();
 
-               expect( mySpy.callCount ).toBe( 0 );
+               expect( mySpy.calls.count() ).toEqual( 0 );
             } );
 
             //////////////////////////////////////////////////////////////////////////////////////////////////
@@ -386,12 +393,12 @@ define( [
                publishFlagChange( 'helpAvailable', true );
                publishFlagChange( 'notUndoable', false );
 
-               testBed_.scope.handleButtonClicked( nextButton );
-               testBed_.scope.handleButtonClicked( helpButton );
-               testBed_.scope.handleButtonClicked( previousButton );
-               jasmine.Clock.tick( 0 );
+               widgetScope.handleButtonClicked( nextButton );
+               widgetScope.handleButtonClicked( helpButton );
+               widgetScope.handleButtonClicked( previousButton );
+               testEventBus.flush();
 
-               expect( mySpy.callCount ).toBe( 3 );
+               expect( mySpy.calls.count() ).toEqual( 3 );
             } );
 
          } );
@@ -408,22 +415,22 @@ define( [
          var nextButton;
 
          beforeEach( function() {
-            nextButton = testBed_.scope.model.areas.right[ 1 ];
+            nextButton = widgetScope.model.areas.right[ 1 ];
 
             var action;
-            testBed_.scope.eventBus.subscribe( 'takeActionRequest', function( event, meta ) {
+            widgetScope.eventBus.subscribe( 'takeActionRequest', function( event, meta ) {
                action = event.action;
                actionEvent = event;
                actionEventName = meta.name;
-               testBed_.scope.eventBus.publish( 'willTakeAction.' + action, { options: { sender: 'spec' } } );
-               jasmine.Clock.tick( 0 );
+               widgetScope.eventBus.publish( 'willTakeAction.' + action, { options: { sender: 'spec' } } );
+               testEventBus.flush();
             } );
             signalActionFinished = function() {
-               testBed_.scope.eventBus.publish( 'didTakeAction.' + action, { options: { sender: 'spec' } } );
-               jasmine.Clock.tick( 0 );
+               widgetScope.eventBus.publish( 'didTakeAction.' + action, { options: { sender: 'spec' } } );
+               testEventBus.flush();
             };
-            testBed_.scope.handleButtonClicked( nextButton );
-            jasmine.Clock.tick( 0 );
+            widgetScope.handleButtonClicked( nextButton );
+            testEventBus.flush();
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -447,16 +454,15 @@ define( [
 
          describe( 'and the action is canceled', function() {
 
-            beforeEach( function() {
+            beforeEach( ngMocks.inject( function( $q ) {
                signalActionFinished();
-               var q = ax.testing.portalMocks.mockQ();
-               testBed_.scope.eventBus.publishAndGatherReplies.andCallFake( q.reject );
-               testBed_.scope.handleButtonClicked( nextButton );
-            } );
+               widgetEventBus.publishAndGatherReplies.and.callFake( $q.reject );
+               widgetScope.handleButtonClicked( nextButton );
+            } ) );
 
             it( 'resets the button state', function() {
                expect( nextButton.classes[ 'ax-active' ] ).toBe( true );
-               jasmine.Clock.tick( 0 );
+               widgetScope.$apply();
                expect( nextButton.classes[ 'ax-active' ] ).toBe( false );
             } );
 
@@ -465,7 +471,7 @@ define( [
          /////////////////////////////////////////////////////////////////////////////////////////////////////
 
          it( 'sends the button\'s id as event.anchorDomElement (R2.9)', function() {
-            expect( actionEvent.anchorDomElement ).toEqual( testBed_.scope.id( 'next_0' ) );
+            expect( actionEvent.anchorDomElement ).toEqual( widgetScope.id( 'next_0' ) );
          } );
 
          /////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -476,13 +482,13 @@ define( [
                expect( actionEventName ).toEqual( 'takeActionRequest.next' );
 
                actionEventName = 'resetEventName';
-               testBed_.scope.handleButtonClicked( nextButton );
-               jasmine.Clock.tick( 0 );
+               widgetScope.handleButtonClicked( nextButton );
+               testEventBus.flush();
                expect( actionEventName ).toEqual( 'resetEventName' );
 
                signalActionFinished();
-               testBed_.scope.handleButtonClicked( nextButton );
-               jasmine.Clock.tick( 0 );
+               widgetScope.handleButtonClicked( nextButton );
+               testEventBus.flush();
                expect( actionEventName ).toEqual( 'takeActionRequest.next' );
             } );
          } );
@@ -494,15 +500,13 @@ define( [
 
    describe( 'An AxCommandBarWidget with buttons with same action', function() {
 
-      beforeEach( function() {
-         testBed_ = testBedAfterDidNavigate( this, testData.customWithSameAction );
-      } );
+      createSetup( testData.customWithSameAction );
 
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       it( 'has individual ids for each button (R2.9)', function() {
-         expect( testBed_.scope.model.areas.left[ 0 ].id ).not.
-            toEqual( testBed_.scope.model.areas.left[ 1 ].id );
+         expect( widgetScope.model.areas.left[ 0 ].id ).not.
+            toEqual( widgetScope.model.areas.left[ 1 ].id );
       } );
 
    } );
@@ -511,28 +515,17 @@ define( [
 
    describe( 'An AxCommandBarWidget', function() {
 
-      function setup( self, layout ) {
+      describe( 'with no explicit layout configuration', function() {
          var features = {
             buttons: [
                { i18nHtmlLabel: 'Action 1', action: 'action1' },
                { i18nHtmlLabel: 'Action 2', action: 'action2' }
             ]
          };
-
-         if( layout ) {
-            features.layout = layout;
-         }
-
-         testBed_ = testBedAfterDidNavigate( self, features );
-      }
-
-      ////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-      describe( 'with no explicit layout configuration', function() {
+         createSetup( features );
 
          it( 'displays the buttons horizontally (R4.1)', function() {
-            setup( this );
-            expect( testBed_.scope.model.layout ).toEqual( 'ax-local-horizontal' );
+            expect( widgetScope.model.layout ).toEqual( 'ax-local-horizontal' );
          } );
 
       } );
@@ -540,10 +533,17 @@ define( [
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       describe( 'with configured layout variant HORIZONTAL', function() {
+         var features = {
+            buttons: [
+               { i18nHtmlLabel: 'Action 1', action: 'action1' },
+               { i18nHtmlLabel: 'Action 2', action: 'action2' }
+            ],
+            layout: { variant: 'HORIZONTAL' }
+         };
+         createSetup( features );
 
          it( 'displays the buttons horizontally (R4.1)', function() {
-            setup( this, { variant: 'HORIZONTAL' } );
-            expect( testBed_.scope.model.layout ).toEqual( 'ax-local-horizontal' );
+            expect( widgetScope.model.layout ).toEqual( 'ax-local-horizontal' );
          } );
 
       } );
@@ -551,10 +551,17 @@ define( [
       ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
       describe( 'with configured layout variant VERTICAL', function() {
+         var features = {
+            buttons: [
+               { i18nHtmlLabel: 'Action 1', action: 'action1' },
+               { i18nHtmlLabel: 'Action 2', action: 'action2' }
+            ],
+            layout: { variant: 'VERTICAL' }
+         };
+         createSetup( features );
 
          it( 'displays the buttons vertically (R4.1)', function() {
-            setup( this, { variant: 'VERTICAL' } );
-            expect( testBed_.scope.model.layout ).toEqual( 'ax-local-vertical' );
+            expect( widgetScope.model.layout ).toEqual( 'ax-local-vertical' );
          } );
 
       } );
@@ -562,46 +569,13 @@ define( [
 
    ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-   function testBedAfterDidNavigate( self, features ) {
-      addButtonMatchers( self );
-
-      var testBed = ax.testing.portalMocksAngular.createControllerTestBed( descriptor );
-      testBed.featuresMock = features;
-
-      testBed.useWidgetJson();
-      testBed.setup();
-
-      return testBed;
-   }
-
-   ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-   function addButtonMatchers( self ) {
-      self.addMatchers( {
-         toContainButtonWithAction: function( expectedAction ) {
-            var buttonList = this.actual;
-            for( var i = 0; i < buttonList.length; ++i ) {
-               if( buttonList[ i ].action === expectedAction ) {
-                  return true;
-               }
-            }
-            return false;
-         },
-         toHaveAction: function( expectedAction ) {
-            return this.actual.action === expectedAction;
-         }
-      } );
-   }
-
-   ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
    function changeLocale( languageTag, locale ) {
       locale = locale || 'default';
-      testBed_.eventBusMock.publish( 'didChangeLocale.' + locale, {
+      testEventBus.publish( 'didChangeLocale.' + locale, {
          locale: locale,
          languageTag: languageTag
       } );
-      jasmine.Clock.tick( 0 );
+      testEventBus.flush();
    }
 
 } );